Move filter menu to it's own component
